// frontend/profdiff-web-application/src/app/models/dto/compilationUnitMetadata.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents a compilation unit with its ID given by HotSpot VM compiler, measured running period, and hot status.
 */
export interface CompilationUnitMetadata {
  /**
   * Unique identifier given by GraalVM compiler.
   */
  id: string;
  /**
   * The number of cycles sampled by perf for the given compilation unit.
   */
  period?: number;
  /**
   * Indication whether the Profdiff marked the compilation unit as \'hot\'.
   */
  isHot?: boolean;
  /**
   * Indication whether the Compilation Unit represents a subtree of a larger, original compilation.
   */
  isFragment: boolean;
}